test(logger): fix logger testing
